Smart Course Material Q&A Bot/requirements.txt
# Core LangChain dependencies
langchain==0.3.26
langchain-core==0.3.66
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
langchain-community>=0.3.0

# Google AI dependencies
google-generativeai>=0.8.0
google-ai-generativelanguage>=0.6.0
google-api-core>=2.19.0
google-auth>=2.23.0

# Document processing
PyPDF2>=3.0.0
pypdf>=4.0.0
python-docx>=1.1.0
unstructured>=0.14.0
python-pptx>=0.6.0
docx2txt>=0.8

# Vector stores and embeddings
faiss-cpu>=1.7.0
numpy>=1.24.0

# Environment and configuration
python-dotenv>=1.0.0

# Web interface
streamlit>=1.28.0

# Utilities
requests>=2.32.0
tqdm>=4.65.0

# Core Python dependencies
pydantic>=2.5.0
pydantic_core>=2.14.0
typing_extensions>=4.8.0
packaging>=23.0
PyYAML>=6.0.0

# HTTP and async support
httpx>=0.25.0
anyio>=4.0.0
tenacity>=8.2.0

# Additional utilities
jsonpatch>=1.33
protobuf>=4.21.0
# Essential dependencies for Smart Grid RAG System
python-dotenv
google-generativeai
langchain
langchain-google-genai  
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
streamlit
requests
PyPDF2
python-docx
python-pptx
docx2txt
pandas
numpy
pydantic
